requirements.txt
langgraph>=0.4.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-tavily>=0.1.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
streamlit>=1.40.0
google-genai>=1.0
Pillow>=10.0
requests>=2.31.0
openai>=1.0
